test: add bitmap rendering tests and split BitmapInfoHeaderTests
Add tests for BitmapInfo.swift rendering functions (sampled32, sampled24,
sampled16, indexed8) to improve coverage. Split BitmapInfoHeaderTests into
separate BitmapRenderTests.swift file to comply with SwiftLint file_length
rule.
