+++ WhiskyKit/Tests/WhiskyKitTests/BinaryParsingTests.swift
@@ -126,147 +126,6 @@ final class ColorQuadTests: XCTestCase {
     }
 }
 
-// MARK: - BitmapInfoHeader Tests
-
-/// Parameters for creating test bitmap info header data
-struct BitmapHeaderParams {
-    var size: UInt32 = 40
-    var width: Int32 = 32
-    var height: Int32 = 32
-    var planes: UInt16 = 1
-    var bitCount: UInt16 = 32
-    var compression: UInt32 = 0
-    var sizeImage: UInt32 = 0
-}
-
-final class BitmapInfoHeaderTests: XCTestCase {
-    var tempDir: URL!
-
-    override func setUp() {
-        super.setUp()
-        tempDir = FileManager.default.temporaryDirectory.appending(path: "bitmap_test_\(UUID().uuidString)")
-        try? FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-    }
-
-    override func tearDown() {
-        try? FileManager.default.removeItem(at: tempDir)
-        super.tearDown()
-    }
-
-    func testBitmapInfoHeaderParsing() throws {
-        let headerData = createBitmapInfoHeaderData(BitmapHeaderParams(
-            width: 32,
-            height: 64,
-            sizeImage: 8_192
-        ))
-
-        let fileURL = tempDir.appending(path: "header.bin")
-        try headerData.write(to: fileURL)
-
-        let handle = try FileHandle(forReadingFrom: fileURL)
-        defer { try? handle.close() }
-
-        let header = BitmapInfoHeader(handle: handle, offset: 0)
-
-        XCTAssertEqual(header.size, 40)
-        XCTAssertEqual(header.width, 32)
-        XCTAssertEqual(header.height, 64)
-        XCTAssertEqual(header.planes, 1)
-        XCTAssertEqual(header.bitCount, 32)
-        XCTAssertEqual(header.compression, .rgb)
-        XCTAssertEqual(header.colorFormat, .sampled32)
-        XCTAssertEqual(header.originDirection, .bottomLeft)
-    }
-
-    func testBitmapInfoHeaderNegativeHeight() throws {
-        let headerData = createBitmapInfoHeaderData(BitmapHeaderParams(
-            height: -64,
-            bitCount: 24,
-            sizeImage: 6_144
-        ))
-
-        let fileURL = tempDir.appending(path: "negheight.bin")
-        try headerData.write(to: fileURL)
-
-        let handle = try FileHandle(forReadingFrom: fileURL)
-        defer { try? handle.close() }
-
-        let header = BitmapInfoHeader(handle: handle, offset: 0)
-
-        XCTAssertEqual(header.height, -64)
-        XCTAssertEqual(header.originDirection, .upperLeft)
-    }
-
-    func testBitmapInfoHeaderDifferentBitCounts() throws {
-        let testCases: [(bitCount: UInt16, expectedFormat: ColorFormat)] = [
-            (1, .indexed1),
-            (4, .indexed4),
-            (8, .indexed8),
-            (16, .sampled16),
-            (24, .sampled24),
-            (32, .sampled32)
-        ]
-
-        for testCase in testCases {
-            let headerData = createBitmapInfoHeaderData(BitmapHeaderParams(
-                width: 16,
-                height: 16,
-                bitCount: testCase.bitCount
-            ))
-
-            let fileURL = tempDir.appending(path: "bc\(testCase.bitCount).bin")
-            try headerData.write(to: fileURL)
-
-            let handle = try FileHandle(forReadingFrom: fileURL)
-            defer { try? handle.close() }
-
-            let header = BitmapInfoHeader(handle: handle, offset: 0)
-
-            XCTAssertEqual(
-                header.colorFormat,
-                testCase.expectedFormat,
-                "Expected \(testCase.expectedFormat) for bitCount \(testCase.bitCount)"
-            )
-        }
-    }
-
-    func testBitmapInfoHeaderHashable() throws {
-        let headerData = createBitmapInfoHeaderData(BitmapHeaderParams(
-            sizeImage: 4_096
-        ))
-
-        let fileURL = tempDir.appending(path: "hash.bin")
-        try headerData.write(to: fileURL)
-
-        let handle = try FileHandle(forReadingFrom: fileURL)
-        defer { try? handle.close() }
-
-        let header1 = BitmapInfoHeader(handle: handle, offset: 0)
-        let header2 = BitmapInfoHeader(handle: handle, offset: 0)
-
-        XCTAssertEqual(header1, header2)
-        XCTAssertEqual(header1.hashValue, header2.hashValue)
-    }
-
-    private func createBitmapInfoHeaderData(_ params: BitmapHeaderParams) -> Data {
-        var data = Data()
-
-        data.append(contentsOf: withUnsafeBytes(of: params.size.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.width.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.height.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.planes.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.bitCount.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.compression.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: params.sizeImage.littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: Int32(0).littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: Int32(0).littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: UInt32(0).littleEndian) { Array($0) })
-        data.append(contentsOf: withUnsafeBytes(of: UInt32(0).littleEndian) { Array($0) })
-
-        return data
-    }
-}
-
 // MARK: - LinkFlags Tests
 
 final class LinkFlagsTests: XCTestCase {
